+++ vllm_ascend/worker/v2/spec_decode/multi_stage/acceptance.py
@@ -43,19 +43,22 @@ def assemble_verified_tokens(
     """
     if steps is None:
         steps = torch.arange(num_speculative_steps + 1, device=logits.device)
+    # MRV2 supplies int32 boundaries. Normalize once instead of implicitly
+    # promoting them again in each index/mask expression below.
+    cu_num_logits = cu_num_logits.long()
     lengths = cu_num_logits[1:] - cu_num_logits[:-1]
     if policy.method == "all":
         # Every request terminates at its bonus row. No vocabulary scan,
         # acceptance mask or first-rejection reduction is needed.
         stop = lengths - 1
-        rows = (cu_num_logits[:-1, None] + steps + 1).clamp(max=draft_sampled.shape[0] - 1).long()
-        bonus = target_sampled[(cu_num_logits[1:] - 1).long()]
+        rows = (cu_num_logits[:-1, None] + steps + 1).clamp(max=draft_sampled.shape[0] - 1)
+        bonus = target_sampled[cu_num_logits[1:] - 1]
         sampled = torch.where(steps < stop[:, None], draft_sampled[rows], bonus[:, None])
         sampled.masked_fill_(steps > stop[:, None], -1)
         return sampled.long(), lengths.to(torch.int32)
     draft_next = draft_sampled.roll(-1)
     accepted = policy.accept(logits, draft_next)
-    rows = (cu_num_logits[:-1, None] + steps).clamp(max=logits.shape[0] - 1).long()
+    rows = (cu_num_logits[:-1, None] + steps).clamp(max=logits.shape[0] - 1)
     is_draft = steps < lengths[:, None] - 1
     # Bonus positions also terminate the prefix, including requests with no draft.
     stop = torch.where(is_draft & accepted[rows], num_speculative_steps + 1, steps).amin(dim=1)
